moved to m_list instead of sys/queue

+++ src/util/count.h
@@ -4,7 +4,7 @@
 #include <sys/stdint.h>
 
 /* Type used everywhere where count (number of) some items is relevant. */
-typedef uint32_t ctf_count;
+typedef uint64_t ctf_count;
 #define CTF_COUNT_MIN          0
 #define CTF_COUNT_MAX 0xffffffff
